+++ repository/user.go
@@ -10,7 +10,7 @@ import (
 
 type UserRepository interface {
 	Create(ctx context.Context, user models.User) error
-	Get(ctx context.Context, id uint) (models.User, error)
+	Get(ctx context.Context, id uint64) (models.User, error)
 	GetByEmail(ctx context.Context, email string) (models.User, error)
 	GetByPhone(ctx context.Context, phone string) (models.User, error)
 }
@@ -27,7 +27,7 @@ func (r user) Create(ctx context.Context, user models.User) error {
 	return r.DB.Create(&user).Error
 }
 
-func (r user) Get(ctx context.Context, id uint) (models.User, error) {
+func (r user) Get(ctx context.Context, id uint64) (models.User, error) {
 	var u models.User
 	err := r.DB.First(&u, id).Error
 	return u, err
